preliminary five stage CPU working
Re-arranged five cycle MIPS CPU so that there is a module
for each stage as opposed to each module having multiple
stages for its outputs.  This design is simpler and easier
to understand.

The five stage CPU appears to be working using the basic
test (01-no_hazard/) which contains no hazards.
Currently the test bench works by outputting various
important variables that change during the cycle.  These
results can then be compared to what is expected (main_tb.check).
But an automated way of generating all the correct results
has not been devised so initially each case must be manually
checked.  It is however quick to test if a new change produces
different results compared to the expected results (main_tb.check).