basic five stage CPU working in Verilog and Logisim
The basic five stage CPU is working but it does not
avoid hazards and jumps aren't supported yet.

In addition to a Verilog version of the five stage MIPS CPU
a Logisim version is also provided.

Changed Verilog module arguments to all be placed at the
top along with their input specifications as opposed
to names at the top with specifications scattered throughout
the source.