refactored cpu to eliminate "stage" modules
Previously there was a module for each stage.
But managing all the wires in to and out of each
module was found to been tedious and cumbersome.

This has been changed so the internals of each stage
are in the cpu and all are at the same scope.
This makes it easy to connect wires to any other
point in the various stages.

The "reggy" module is also used to register
variables as the pass from one stage to the next.

All changes have been tested against the current
test benches and no problems have been detected.