fixed forwarding, implemented stalling
The previous implementation of forwarding could forward
from the data memory output.  This was bad because it
would increase the critical path which would require a
slower clock.

This has been fixed by adding the ability to detect
a load use data hazard in the decode stage and insert
a nop (bubble).  The nop is created by setting all the
control signals to zero at the ID/EX register.