Fixed a sw, lw, stall bug
There was a bug when a sw is performed immediately after
a lw, the current value would not be saved.  This was caused
by the value of data2 which is passed to stage4 not coming
from a forwarded location.

The check files of all the test benches have been updated
for a recent change in how the instruction memory is registered.
No differences were found in any of the results.