further testing of lw sw hazard
Created additional test benches for testing the lw sw
scenarios which can result in load use and read after
write hazards.  All situations appear to be handled correctly.

The test bench '07-lw_sw' is different from previous
tests in that it is simpler to check.
It performs a bunch of additions which should result
in a specific answer at the end if nothing went wrong.
And this result can be easily verified by hand calculations.
This is much less tedious than going through every
operation in the test bench and checking that every result
is correct at every step.