assemble using Gcc
Previously, the test code was assembled using a custom assembler which
was not included with this project.  This made it difficult to try out
new code when the assembler was not available.  Improve this situation
by assembling the code using a standard Gcc MIPS assembler.

Compilation was tested using the MIPS cross compiler available in
Debian (binutils-mips-linux-gnu).  An additional program, bin2hex.c, was
created to convert the binary dump to ASCII hex format suitable for use
with a Verilog simulator.