rebuild test/*.hex using Gcc, branch with noop
For whatever reason Gcc adds a noop after some branches.  This should
make no difference.  But it also causes a change in the branch opcode
since the destination address is one more than it was previously.

Re-assemble some of test/*.hex using Gcc, those which had a branch
instruction where a noop was added.