simplify using `DEBUG_CPU_STAGES for debug output
Previously, the debug output for each stage in the cpu (cpu.v) was
brought out through the interface using wires in to the cpu test bench
(cpu_tb.v) where $display was called.

Simplify the design by creating a `DEBUG_CPU_STAGES define for turning
the output on.  And place the $display of this output in cpu.v where
the variables are immediately available.  This eliminates the need to
run the wires through the cpu.v interface.

Placing the $display statement in cpu.v caused the $dumpvars output to
come after the header which changed the .out files.  Updated all the
.check files by swapping the first two header lines.