add final value capability to test benches
All the previous test benches worked by dumping a select set of
variables in each stage of the processor.  This produced a large dump of
the output which could then be diffed against a known good dump (.check)
to see if there were any errors.  However, the tiniest of change will
create many changes in the dump, which may or may not be significant.
And any change requires tedious re-examination to verify that everything
is still correct.

Add a simplified test bench which only tests the final values of the
registers.