add support for bne (branch not equal) operation
Previously there was no support for bne (branch not equal).  The opcode
was not checked control.v did not output appropriate control signals.

Support was added by expanding the branch signal from 1 bit to 2 bits so
it can now be tested whether the branch is for a bne or beq.