refactor control.v to be logical and simple
Previously, control.v had two sets of variables.  One were
flags based on the opcode.  And the second set were flags
based on the first set of flags.  But this made determining
what opcode did what somewhat confusing.

Refactor the code in to one large case statement so that
it is easier to understand what each opcode is doing.